ordermanagementsystem/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.0
astroid==2.11.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
base58==2.1.1
bitarray==2.5.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
cytoolz==0.11.2
Deprecated==1.2.13
dill==0.3.4
distlib==0.3.4
eip712-structs==1.1.0
eth-abi==2.1.1
eth-account==0.5.8
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
filelock==3.7.0
Flask==2.1.2
frozenlist==1.3.0
grpc-tools==0.0.1
grpcio==1.46.3
grpcio-tools==1.46.3
hexbytes==0.2.2
html5lib==1.1
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
keyring==23.5.1
lazy-object-proxy==1.7.1
lockfile==0.12.2
lru-dict==1.1.7
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mypy==0.950
mypy-extensions==0.4.3
netaddr==0.8.0
packaging==20.9
parsimonious==0.8.1
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.5.2
pluggy==1.0.0
poetry-core==1.0.8
protobuf==3.20.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
py-clob-client==0.0.22
py-order-utils==0.0.21
pycodestyle==2.8.0
pycryptodome==3.14.1
pylev==1.4.0
pylint==2.13.9
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
pysha3==1.0.2
pytest==7.1.2
python-dotenv==0.20.0
redis==4.3.1
requests==2.27.1
requests-toolbelt==0.9.1
rlp==2.0.1
shellingham==1.4.0
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
toolz==0.11.2
types-redis==4.2.5
types-requests==2.27.27
types-urllib3==1.26.14
typing_extensions==4.2.0
urllib3==1.26.9
varint==1.0.2
virtualenv==20.14.1
web3==5.29.2
webencodings==0.5.1
websocket-client==1.3.2
websockets==9.1
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
